Convert createSexuality to TS (#462)
